backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
python-socketio==5.16.1
litellm==1.83.3
SQLAlchemy==2.0.49
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.0.1
aiohttp==3.13.5
